bugs fix: 1. heuristic return number of save people (including the number of people the other agent carrying and able to save- that's why hurestic get the states of both agent 1 and agent 2) 2. update currnode after agent decided (we forgot and that's why in agent position never changed in UI) 3. update game state after agent decided 4.

+++ src/main/java/states/GameState.java
@@ -10,7 +10,7 @@ import java.util.Map;
 public class GameState extends State {
 
     @Getter @Setter
-    private double mineSavedPeople;
+    private int mineSavedPeople;
 
     @Getter @Setter
     private HurricaneNode otherCurrNode;
@@ -21,7 +21,7 @@ public class GameState extends State {
     @Getter @Setter
     private int otherPeople;
 
-    public GameState(SimulatorContext simulatorContext, HurricaneNode currNode, State prev, int people, double costSoFar, double savedPeople, HurricaneNode otherCurrNode, double otherSavedPeople, int otherPeople) {
+    public GameState(SimulatorContext simulatorContext, HurricaneNode currNode, State prev, int people, double costSoFar, int savedPeople, HurricaneNode otherCurrNode, double otherSavedPeople, int otherPeople) {
         super(simulatorContext, currNode, prev, people, costSoFar);
         this.time = simulatorContext.getTime();
         this.mineSavedPeople = savedPeople;
@@ -36,7 +36,7 @@ public class GameState extends State {
                      double time,
                      int people,
                      double costSoFar,
-                     double savedPeople,
+                     int savedPeople,
                      HurricaneNode otherCurrNode,
                      double otherSavedPeople,
                      int otherPeople) {
